Fixed fileParser in order to read time scale in decimal
- json-simple can read either long or double. Updated getScaleTime to make it capable to read a decimal number.

#fix #story[5]

+++ src/main/java/seng302/App.java
@@ -20,7 +20,7 @@ public class App
 		int numberOfBoats = (int) fp.getRaceSize();
 
 		//get time scale
-		int timeScale = (int) fp.getTimeScale();
+		double timeScale = fp.getTimeScale();
 		race.setTimeScale(timeScale);
 
 		for (Map<String, Object> team : teams) {
